\name{Lam07.Summary}
\alias{Lam07.Summary}
\docType{data}
\title{
Summary network meta-analysis results of network Lam 2007
}
\description{
\code{Lam07.Summary} is a summary results for network Lam 2007 given by the function \code{\link{nma.ab}}. This is served as an example for function \code{\link{ci.plot}} and \code{\link{create.tab}} in this \pkg{pcnetmeta} package. These results are generated by \code{\link{nma.ab}} using the homogeneous model, setting \code{n.iter} as 1,000,000 and \code{n.chains} as 2. Also, it contains nodes "probt", "RR", "RD", "OR", "rk", "best", "mu", "sigma", "vi".
}
\usage{data(Lam07.Summary)}
\format{
  A data frame with 93 observations on the following 9 variables. Each row is a summary result of "RR", "RD", "OR", "probt" (estimated event rate) etc.
  \describe{
    \item{\code{mean}}{a numeric vector indicating mean value.}
    \item{\code{sd}}{a numeric vector indicating standard deviation.}
    \item{\code{X2.5.}}{a numeric vector indicating 2.5\% percentile.}
    \item{\code{X25.}}{a numeric vector indicating 25\% percentile.}
    \item{\code{X50.}}{a numeric vector indicating median (50\% percentile).}
    \item{\code{X75.}}{a numeric vector indicating 75\% percentile.}
    \item{\code{X97.5.}}{a numeric vector indicating 97.5\% percentile.}
  }
}
